cleanup after merge and split tests
