+++ src/QirRuntime/lib/QIR/quantum__qis.hpp
@@ -75,11 +75,6 @@ extern "C"
 
     QIR_SHARED_API double      quantum__qis__ieeeremainder__body(double x, double y);       // NOLINT
     QIR_SHARED_API int64_t     quantum__qis__drawrandomint__body(int64_t minimum, int64_t maximum);     // NOLINT
-    QIR_SHARED_API bool quantum__qis__isnan__body(double d);                                    // NOLINT
-    QIR_SHARED_API double quantum__qis__infinity__body();                                       // NOLINT
-    QIR_SHARED_API bool quantum__qis__isinf__body(double d);                                    // NOLINT
-    QIR_SHARED_API double quantum__qis__arctan2__body(double y, double x);                      // NOLINT
-    QIR_SHARED_API int64_t quantum__qis__drawrandomint__body(int64_t minimum, int64_t maximum); // NOLINT
 
     // Q# ApplyIf:
     QIR_SHARED_API void unexpected_conditional();                                                      // NOLINT
