+++ src/QirRuntime/test/QIR-static/qir-driver.cpp
@@ -329,72 +329,3 @@ TEST_CASE("QIR: application of nested controlled functor", "[qir][qir.functor]")
 
     g_ctrqapi = nullptr;
 }
-
-// TestConditionalOnResult() is authored in a way that the expected path through the function only applies X operator
-// for the chosen sequence of measurement results, and all other paths apply Y. Thus, the correct execution must get the
-// expected maximum number of X and ControlledX callback.
-struct ConditionalsTestSimulator : public Microsoft::Quantum::SimulatorStub
-{
-    int cX = 0;
-    int ccX = 0;
-    vector<ResultValue> mockMeasurements;
-
-    explicit ConditionalsTestSimulator(vector<ResultValue>&& results)
-        : mockMeasurements(results)
-    {
-    }
-
-    Qubit AllocateQubit() override
-    {
-        return nullptr;
-    }
-    void ReleaseQubit(Qubit qubit) override {}
-
-    void X(Qubit) override
-    {
-        this->cX++;
-    }
-    void ControlledX(long numControls, Qubit controls[], Qubit qubit) override
-    {
-        this->ccX++;
-    }
-    void Y(Qubit) override {}
-
-    Result Measure(long numBases, PauliId bases[], long numTargets, Qubit targets[]) override
-    {
-        static int callNumber = -1;
-        callNumber++;
-        assert(callNumber < this->mockMeasurements.size() && "ConditionalsTestSimulator isn't set up correctly");
-
-        return (this->mockMeasurements[callNumber] == Result_Zero) ? UseZero() : UseOne();
-    }
-
-    bool AreEqualResults(Result r1, Result r2) override
-    {
-        // those are bogus pointers but it's ok to compare them _as pointers_
-        return (r1 == r2);
-    }
-
-    void ReleaseResult(Result result) override {} // the results aren't allocated by this test simulator
-
-    Result UseZero() override
-    {
-        return reinterpret_cast<Result>(0);
-    }
-
-    Result UseOne() override
-    {
-        return reinterpret_cast<Result>(1);
-    }
-};
-extern "C" void Microsoft__Quantum__Testing__QIR__TestConditionalOnResult__body(); // NOLINT
-TEST_CASE("QIR: conditionals on measurement results", "[qir][qir.conditionals]")
-{
-    unique_ptr<ConditionalsTestSimulator> qapi =
-        make_unique<ConditionalsTestSimulator>(vector<ResultValue>{Result_Zero, Result_One});
-    QirContextScope qirctx(qapi.get(), true /*trackAllocatedObjects*/);
-
-    CHECK_NOTHROW(Microsoft__Quantum__Testing__QIR__TestConditionalOnResult__body());
-    CHECK(qapi->cX == 14);
-    CHECK(qapi->ccX == 9);
-}
