+++ src/QirRuntime/test/QIR-static/qsharp/qir-test-conditionals.qs
@@ -4,7 +4,7 @@ namespace Microsoft.Quantum.Testing.QIR {
     open Microsoft.Quantum.Intrinsic;
     open Microsoft.Quantum.Canon;
 
-    operation TestConditionalOnResult() : Unit {
+    operation TestApplyIf() : Unit {
         use q1 = Qubit();
         use q2 = Qubit();
 
@@ -13,25 +13,12 @@ namespace Microsoft.Quantum.Testing.QIR {
         ApplyIfElseRCA(r1, (X, q1), (Y, q1));
         Adjoint ApplyIfElseRCA(r1, (X, q1), (Y, q1));
         Controlled ApplyIfElseRCA([q2], (r1, (X, q1), (Y, q1)));
+        X(q2);
         Adjoint Controlled ApplyIfElseRCA([q2], (r1, (X, q1), (Y, q1)));
 
         // One branch
-        let r2 = M(q1);
+        let r2 = M(q2);
         ApplyIfElseRCA(r2, (Y, q1), (X, q1));
-
-        // Multiple ops, guarded by the condition
-        if r1 == Zero {
-            X(q1);
-            X(q2);
-        }
-        else {
-            Y(q1);
-        }
-
-        // Multi-result comparison
-        Microsoft.Quantum.Simulation.QuantumProcessor.Extensions.ApplyConditionally([r1], [r2], (Y, q1), (X, q1));
-        Microsoft.Quantum.Simulation.QuantumProcessor.Extensions.ApplyConditionally(
-            [r1, One], [Zero, r2], (X, q1), (Y, q1));
 
         // Other variants
         ApplyIfElseR(r1, (X, q1), (Y, q1));
@@ -47,12 +34,39 @@ namespace Microsoft.Quantum.Testing.QIR {
         Adjoint ApplyIfZeroA(r1, (X, q1));
         Controlled ApplyIfZeroC([q2], (r1, (X, q1)));
         Adjoint Controlled ApplyIfZeroCA([q2], (r1, (X, q1)));
+    }
 
+    operation TestApplyConditionally() : Unit {
+        use q1 = Qubit();
+        use q2 = Qubit();
+
+        let r1 = M(q1);
+        X(q2);
+        let r2 = M(q2);
+
+        Microsoft.Quantum.Simulation.QuantumProcessor.Extensions.ApplyConditionally([r1], [r2], (Y, q1), (X, q1));
+        Microsoft.Quantum.Simulation.QuantumProcessor.Extensions.ApplyConditionally(
+            [r1, One], [Zero, r2], (X, q1), (Y, q1));
+
+        // Other variants
         Adjoint Microsoft.Quantum.Simulation.QuantumProcessor.Extensions.ApplyConditionallyA(
             [r1], [r2], (Y, q1), (X, q1));
         Controlled Microsoft.Quantum.Simulation.QuantumProcessor.Extensions.ApplyConditionallyC(
             [q2], ([r1], [r2], (Y, q1), (X, q1)));
         Adjoint Controlled Microsoft.Quantum.Simulation.QuantumProcessor.Extensions.ApplyConditionallyCA(
             [q2], ([r1], [r2], (Y, q1), (X, q1)));
+    }
+
+    operation TestConditionalRewrite() : Unit {
+        use q1 = Qubit();
+        use q2 = Qubit();
+
+        if M(q1) == Zero {
+            X(q1);
+            Controlled X([q2], q1);
+        }
+        else {
+            Y(q1);
+        }
     }
 }
